@@ -1,16 +1,14 @@
-import { AlertService } from './alert.service';
 import { Injectable, EventEmitter } from '@angular/core';
-// import { Http, Response, Headers, RequestOptions } from '@angular/http';
-
 import { HttpClient, HttpHeaders, HttpRequest, HttpResponse, HttpErrorResponse } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+
 import { environment } from '../../environments/environment';
+import { AlertService } from './alert.service';
 import { Feature } from '../feature';
 import { User } from '../_models/user';
 import { DebugService } from './../_services/debug.service';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import { ErrorObservable } from 'rxjs/observable/ErrorObservable';
+
+import { map, catchError } from 'rxjs/operators';
+import { throwError } from 'rxjs';
 
 @Injectable()
 export class ApiService {
@@ -22,23 +20,15 @@ export class ApiService {
   userUrl = 'https://' + environment.API_URL + '/users/';
   partSubsUrl = `https://${environment.API_URL}/parts_substitutes`;
 
-  constructor(
-    private http: HttpClient,
-    private feature: Feature,
-    private user: User,
-    private debug: DebugService,
-    private alert: AlertService
-  ) { }
+  constructor(private http: HttpClient, private feature: Feature, private user: User, private debug: DebugService, private alert: AlertService) {}
 
   getMyDesigns() {
-    return this.http.get(this.apiUrl + 'list/' + this.user.uid)
-      .catch(this.handleError);
+    return this.http.get(this.apiUrl + 'list/' + this.user.uid).pipe(catchError(this.handleError));
   }
 
   getUserRep(uid: number) {
     this.debug.log('api', 'getting user rep');
-    return this.http.get(this.userUrl + uid + '/rep')
-      .catch(this.handleError);
+    return this.http.get(this.userUrl + uid + '/rep').pipe(catchError(this.handleError));
   }
 
   loadDesign(id: number) {
@@ -50,74 +40,82 @@ export class ApiService {
     this.debug.log('api', 'updating design');
     // we can't forget about the hardware...
     this.debug.log('api', this.feature.tiles);
-    if (this.feature.is_quantity_order) { this.prepDataForQtyOrder(); }
+    if (this.feature.is_quantity_order) {
+      this.prepDataForQtyOrder();
+    }
     const patchData = {
-      'id': this.feature.id,
-      'uid': this.user.uid,
-      'feature_type': this.feature.feature_type,
-      'design_name': this.feature.design_name,
-      'project_name': this.feature.project_name,
-      'specifier': this.feature.specifier,
-      'width': this.feature.width || 0,
-      'length': this.feature.length || 0,
-      'units': this.feature.units,
-      'material': this.feature.material,
-      'tile_size': this.feature.tile_size,
-      'tiles': JSON.stringify(this.feature.tiles),
-      'design_data_url': this.feature.design_data_url,
-      'hardware': (!!this.feature.hardware) ? JSON.stringify(this.feature.hardware) : null,
-      'estimated_amount': this.feature.estimated_amount,
-      'services_amount': this.feature.services_amount,
-      'grid_data': JSON.stringify(this.feature.gridData),
-      'quoted': this.feature.quoted,
-      'archived': this.feature.archived,
-      'quantity': this.feature.quantity,
-      'is_quantity_order': this.feature.is_quantity_order
+      id: this.feature.id,
+      uid: this.user.uid,
+      feature_type: this.feature.feature_type,
+      design_name: this.feature.design_name,
+      project_name: this.feature.project_name,
+      specifier: this.feature.specifier,
+      width: this.feature.width || 0,
+      length: this.feature.length || 0,
+      units: this.feature.units,
+      material: this.feature.material,
+      tile_size: this.feature.tile_size,
+      grid_type: this.feature.grid_type,
+      tiles: JSON.stringify(this.feature.tiles),
+      design_data_url: this.feature.design_data_url,
+      hardware: !!this.feature.hardware ? JSON.stringify(this.feature.hardware) : null,
+      estimated_amount: this.feature.estimated_amount,
+      services_amount: this.feature.services_amount,
+      grid_data: JSON.stringify(this.feature.gridData),
+      quoted: this.feature.quoted,
+      archived: this.feature.archived,
+      quantity: this.feature.quantity,
+      is_quantity_order: this.feature.is_quantity_order
     };
 
-    return this.http.patch(this.apiUrl + this.feature.id, patchData)
-      .map((res) => {
+    return this.http.patch(this.apiUrl + this.feature.id, patchData).pipe(
+      map((res: any) => {
         this.onSaved.emit();
         this.debug.log('api', 'emitting onSaved in updateDesign');
-        return res || {}
-      })
-      .catch(this.handleError);
+        return res || {};
+      }),
+      catchError(this.handleError)
+    );
   }
 
   saveDesign() {
     this.debug.log('api', 'saving design');
     const featureType = this.feature.setFeatureType(this.feature.feature_type);
-    if (this.feature.is_quantity_order) { this.prepDataForQtyOrder(); }
-    const patchData = {
-      'uid': this.user.uid,
-      'feature_type': featureType,
-      'design_name': this.feature.design_name,
-      'project_name': this.feature.project_name,
-      'specifier': this.feature.specifier,
-      'width': this.feature.width || 0,
-      'length': this.feature.length || 0,
-      'units': this.feature.units,
-      'material': this.feature.material,
-      'tile_size': this.feature.tile_size,
-      'tiles': JSON.stringify(this.feature.tiles),
-      'design_data_url': this.feature.design_data_url,
-      'hardware': (!!this.feature.hardware) ? JSON.stringify(this.feature.hardware) : null,
-      'estimated_amount': this.feature.estimated_amount,
-      'services_amount': this.feature.services_amount,
-      'grid_data': JSON.stringify(this.feature.gridData),
-      'quoted': this.feature.quoted,
-      'archived': this.feature.archived,
-      'quantity': this.feature.quantity,
-      'is_quantity_order': this.feature.is_quantity_order
+    if (this.feature.is_quantity_order) {
+      this.prepDataForQtyOrder();
     }
+    const patchData = {
+      uid: this.user.uid,
+      feature_type: featureType,
+      design_name: this.feature.design_name,
+      project_name: this.feature.project_name,
+      specifier: this.feature.specifier,
+      width: this.feature.width || 0,
+      length: this.feature.length || 0,
+      units: this.feature.units,
+      material: this.feature.material,
+      tile_size: this.feature.tile_size,
+      grid_type: this.feature.grid_type,
+      tiles: JSON.stringify(this.feature.tiles),
+      design_data_url: this.feature.design_data_url,
+      hardware: !!this.feature.hardware ? JSON.stringify(this.feature.hardware) : null,
+      estimated_amount: this.feature.estimated_amount,
+      services_amount: this.feature.services_amount,
+      grid_data: JSON.stringify(this.feature.gridData),
+      quoted: this.feature.quoted,
+      archived: this.feature.archived,
+      quantity: this.feature.quantity,
+      is_quantity_order: this.feature.is_quantity_order
+    };
 
-    return this.http.post(this.apiUrl, patchData)
-      .map((res: Response) => {
+    return this.http.post(this.apiUrl, patchData).pipe(
+      map((res: any) => {
         this.onSaved.emit();
         this.debug.log('api', 'emitting onSaved in saveDesign');
-        return res || {}
-      })
-      .catch(this.handleError);
+        return res || {};
+      }),
+      catchError(this.handleError)
+    );
   }
 
   prepDataForQtyOrder() {
@@ -130,31 +128,32 @@ export class ApiService {
   }
 
   sendEmail() {
-    return this.http.get(this.apiUrl + 'email/' + this.user.uid + '/design/' + this.feature.id)
+    return this.http.get(this.apiUrl + 'email/' + this.user.uid + '/design/' + this.feature.id);
   }
 
   getPrices() {
-    return this.http.get(this.apiUrl + 'prices')
-      .map((res: Response) => res)
-      .catch(this.handleError);
+    return this.http.get(this.apiUrl + 'prices').pipe(
+      map((res: Response) => res),
+      catchError(this.handleError)
+    );
   }
 
   getPartsSubstitutes() {
-    return this.http.get(this.partSubsUrl)
-      .map((res: Response) => res)
-      .catch(this.handleError);
+    return this.http.get(this.partSubsUrl).pipe(
+      map((res: Response) => res),
+      catchError(this.handleError)
+    );
   }
-
 
   login(email: string, password: string) {
     this.debug.log('api', 'api login');
     const formData = {
-      'email': email,
-      'password': password
-    }
+      email: email,
+      password: password
+    };
 
-    return this.http.post(this.loginUrl, formData)
-      .map((res: any) => {
+    return this.http.post(this.loginUrl, formData).pipe(
+      map((res: any) => {
         console.log('login res:', res);
         if (res && !res.result.error) {
           localStorage.setItem('3formUser', JSON.stringify(res.result.user));
@@ -164,16 +163,17 @@ export class ApiService {
         } else {
           this.alert.apiAlert(res.result.error);
         }
-      })
-      .catch((res) => {
+      }),
+      catchError(res => {
         this.alert.error(res.error.result.message);
         return 'error';
-      });
+      })
+    );
   }
 
   logout() {
     localStorage.removeItem('3formUser');
-    this.user = new User;
+    this.user = new User();
   }
 
   public handleError(error: HttpErrorResponse) {
@@ -186,12 +186,9 @@ export class ApiService {
     } else {
       // The backend returned an unsuccessful response code.
       // The response body may contain clues as to what went wrong,
-      this.debug.log('api',
-        `Backend returned code ${error.status}, body was: ${error.message}`);
+      this.debug.log('api', `Backend returned code ${error.status}, body was: ${error.message}`);
     }
     // return an ErrorObservable with a user-facing error message
-    return new ErrorObservable(
-      'Something bad happened; please try again later.');
-  };
-
+    return throwError('Something bad happened; please try again later.');
+  }
 }
