@@ -1,10 +1,8 @@
 import { Component, OnInit, OnDestroy, AfterContentInit } from '@angular/core';
 import { DesignComponent } from './../design.component';
-import { Subject } from 'rxjs/Subject';
-import 'rxjs/add/operator/takeUntil';
-import 'rxjs/add/operator/filter';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/mergeMap';
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+
 import { NavigationEnd } from '@angular/router/src/events';
 
 @Component({
@@ -25,46 +23,45 @@ export class SeeyondDesignComponent extends DesignComponent implements OnInit, O
   patternRelief: string;
   patternReliefOptions = [
     {
-      'value': 'both',
-      'name': 'front & back'
+      value: 'both',
+      name: 'front & back'
     },
     {
-      'value': 'front',
-      'name': 'front'
+      value: 'front',
+      name: 'front'
     },
     {
-      'value': 'back',
-      'name': 'back'
+      value: 'back',
+      name: 'back'
     }
   ];
 
   ngOnInit() {
-    this.seeyond.onDimensionsChange
-      .takeUntil(this.ngUnsubscribe)
-      .subscribe(data => {
-        this.dimensionsString = this.seeyond.getDimensionString();
-        this.patternRelief = this.getPatternReliefString();
-      });
+    this.seeyond.onDimensionsChange.pipe(takeUntil(this.ngUnsubscribe)).subscribe(data => {
+      this.dimensionsString = this.seeyond.getDimensionString();
+      this.patternRelief = this.getPatternReliefString();
+    });
   }
 
   ngAfterContentInit() {
     // subscribe to the onFeatureUpdated event to update the price.
-    this.seeyond.onFeatureUpdated
-      .takeUntil(this.ngUnsubscribe)
-      .subscribe(data => {
-        this.seeyond.updateEstimatedAmount();
-      });
+    this.seeyond.onFeatureUpdated.pipe(takeUntil(this.ngUnsubscribe)).subscribe(data => {
+      this.seeyond.updateEstimatedAmount();
+    });
 
-    this.seeyond.$outdatedMaterial
-      .takeUntil(this.ngUnsubscribe)
-      .subscribe(data => {
-        if (this.seeyond.materialObj.status === 'inactive') { this.alert.error(`The color \"${this.seeyond.materialObj.name_str}\" is being discontinued.  It will be available until ${this.seeyond.materialObj.available_until} or while supplies last.`)}
-        if (this.seeyond.materialObj.status === 'discontinued') {
-          this.alert.error(`The color \"${this.seeyond.materialObj.name_str}\" has been discontinued.  Select a new color to proceed.`);
-          this.feature.canQuote = false;
-        }
-      });
-
+    this.seeyond.$outdatedMaterial.pipe(takeUntil(this.ngUnsubscribe)).subscribe(data => {
+      if (this.seeyond.materialObj.status === 'inactive') {
+        this.alert.error(
+          `The color \"${this.seeyond.materialObj.name_str}\" is being discontinued.  It will be available until ${
+            this.seeyond.materialObj.available_until
+          } or while supplies last.`
+        );
+      }
+      if (this.seeyond.materialObj.status === 'discontinued') {
+        this.alert.error(`The color \"${this.seeyond.materialObj.name_str}\" has been discontinued.  Select a new color to proceed.`);
+        this.feature.canQuote = false;
+      }
+    });
   }
 
   ngOnDestroy() {
@@ -77,7 +74,10 @@ export class SeeyondDesignComponent extends DesignComponent implements OnInit, O
   }
 
   public updateSelectedTessellation(tessellationName: string) {
-    if (this.seeyond.quoted) { this.alertQuoted(); return; }
+    if (this.seeyond.quoted) {
+      this.alertQuoted();
+      return;
+    }
     this.seeyond.tessellationStr = tessellationName;
     const tessellation = this.seeyond.getTesslationNumber(tessellationName);
     this.selectedTessellation = this.seeyond.tessellation = tessellation;
@@ -87,7 +87,10 @@ export class SeeyondDesignComponent extends DesignComponent implements OnInit, O
   }
 
   public updateSelectedMaterial(material) {
-    if (this.seeyond.quoted) { this.alertQuoted(); return; }
+    if (this.seeyond.quoted) {
+      this.alertQuoted();
+      return;
+    }
     this.selectedMaterial = this.seeyond.material = material.material;
     this.seeyond.sheet_part_id = material.sheet_part_id;
     this.seeyond.canQuote = true;
@@ -104,14 +107,20 @@ export class SeeyondDesignComponent extends DesignComponent implements OnInit, O
   }
 
   public updateUnits(units) {
-    if (this.seeyond.quoted) { this.alertQuoted(); return; }
+    if (this.seeyond.quoted) {
+      this.alertQuoted();
+      return;
+    }
     this.seeyond.units = units;
     this.seeyond.convertDimensionsUnits(units);
     this.dimensionsString = this.seeyond.getDimensionString(units);
   }
 
   public toggleCoveLighting() {
-    if (this.seeyond.quoted) { this.alertQuoted(); return; }
+    if (this.seeyond.quoted) {
+      this.alertQuoted();
+      return;
+    }
     this.seeyond.cove_lighting = !this.seeyond.cove_lighting;
     if (this.seeyond.cove_lighting) {
       this.seeyond.calcLightingFootage();
@@ -124,17 +133,17 @@ export class SeeyondDesignComponent extends DesignComponent implements OnInit, O
       case 'front':
         this.seeyond.front_relief = true;
         this.seeyond.back_relief = false;
-      break;
+        break;
 
       case 'back':
         this.seeyond.front_relief = false;
         this.seeyond.back_relief = true;
-      break;
+        break;
 
       default:
         this.seeyond.front_relief = true;
         this.seeyond.back_relief = true;
-      break;
+        break;
     }
 
     this.debug.log('seeyond-design', this.seeyond.front_relief);
