@@ -1,8 +1,8 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-@Pipe({ name: 'keys',  pure: false })
+@Pipe({ name: 'keys', pure: false })
 export class KeysPipe implements PipeTransform {
-    transform(value: any, args: any[] = null): any {
-        return Object.keys(value).map(key => value[key]);
-    }
+  transform(value: any, args: any[] = null): any {
+    return Object.keys(value).map(key => value[key]);
+  }
 }
