@@ -5,11 +5,9 @@ import * as debug from 'debug';
 export class DebugService {
   private debug = debug;
 
-  constructor() {
-  }
+  constructor() {}
 
   log(identifier: string, message: any) {
     this.debug('ceilings:' + identifier)(message);
   }
-
 }
