@@ -10,19 +10,15 @@ import { AlertService } from '../_services/alert.service';
 export class AlertComponent implements OnInit {
   message: any;
 
-  constructor(
-    private alert: AlertService,
-    public snackBar: MatSnackBar
-  ) { }
+  constructor(private alert: AlertService, public snackBar: MatSnackBar) {}
 
   ngOnInit() {
     this.alert.getMessage().subscribe(message => {
       const config = new MatSnackBarConfig();
-      config.extraClasses = message.type === 'error' ? ['ceilings-alert-error'] : ['ceilings-alert-success'];
+      config.panelClass = message.type === 'error' ? ['ceilings-alert-error'] : ['ceilings-alert-success'];
       config.duration = message.type === 'success' ? 1000 : null;
       config.announcementMessage = message.text;
       this.snackBar.open(message.text, 'dismiss', config);
     });
   }
-
 }
