@@ -4,6 +4,7 @@ import { Injectable, EventEmitter } from '@angular/core';
 import { DebugService } from './_services/debug.service';
 import * as _ from 'lodash';
 import { Location } from '@angular/common';
+import { GridSection } from './_models/grid-section';
 
 @Injectable()
 export class Feature {
@@ -28,8 +29,8 @@ export class Feature {
   public tiles: any;
   public design_data_url: any;
   public hardware: any;
-  public estimated_amount = 0.00;
-  public services_amount = 0.00;
+  public estimated_amount = 0.0;
+  public services_amount = 0.0;
   public front_relief = true; // boolean
   public back_relief = false; // boolean
   public quoted = false; // boolean
@@ -39,10 +40,12 @@ export class Feature {
   public is_quantity_order = false;
   public qtyTilesUsed = 0;
   public qtyTilesReceiving = 0;
+  public grid_type: string = undefined;
 
   // attributes for the tool
   public tile_type = 'tile';
-  public selectedTile = '01';
+  public selectedTile: any;
+  public tile_image_type = 24;
   public selectedTool: string;
   public showGuide = true;
   public materialHex: string;
@@ -51,6 +54,7 @@ export class Feature {
   public discontinuedMaterials: Array<string>;
   public inactiveMaterials: Array<string>;
   public canQuote = true;
+  public clairoTileSizeType = 'standard';
 
   public gridData: any;
   public toolsArray = this.materialsService.toolsArray;
@@ -59,12 +63,7 @@ export class Feature {
   public materialObj: any;
   public seeyond_features = this.materialsService.seeyond_features;
 
-  constructor(
-    public materialsService: MaterialsService,
-    public debug: DebugService,
-    public location: Location,
-    public alert: AlertService
-  ) {}
+  constructor(public materialsService: MaterialsService, public debug: DebugService, public location: Location, public alert: AlertService) {}
 
   setDesign(design: any) {
     this.id = design.id;
@@ -80,7 +79,7 @@ export class Feature {
     this.tile_size = design.tile_size;
     this.design_data_url = design.design_data_url;
     this.tiles = JSON.parse(design.tiles);
-    this.hardware = (!!design.hardware) ? JSON.parse(design.hardware) : null;
+    this.hardware = !!design.hardware ? JSON.parse(design.hardware) : null;
     this.estimated_amount = design.estimated_amount;
     this.services_amount = design.services_amount;
     this.gridData = JSON.parse(design.grid_data);
@@ -114,12 +113,12 @@ export class Feature {
     this.length = undefined;
     this.units = 'inches';
     this.material = undefined;
-    this.tile_size = 24;
+    this.tile_image_type = 24;
     this.tiles = undefined;
     this.design_data_url = undefined;
     this.hardware = undefined;
-    this.estimated_amount = 0.00;
-    this.services_amount = 0.00;
+    this.estimated_amount = 0.0;
+    this.services_amount = 0.0;
     this.front_relief = true; // boolean
     this.back_relief = false; // boolean
     this.quoted = false; // boolean
@@ -148,7 +147,9 @@ export class Feature {
   }
 
   getDeprecatedMaterials() {
-    if (this.feature_type === 'seeyond') { return; } // handled by seeyondFeature.loadSeeyondDesign()
+    if (this.feature_type === 'seeyond') {
+      return; // handled by seeyondFeature.loadSeeyondDesign()
+    }
     const inactiveMaterials = [];
     const discontinuedMaterials = [];
     const materialsObj = this.materials;
@@ -159,10 +160,10 @@ export class Feature {
             for (const matTypeColor in materialsObj[mat][matType]) {
               if (materialsObj[mat][matType].hasOwnProperty(matTypeColor)) {
                 if (materialsObj[mat][matType][matTypeColor].status === 'inactive') {
-                  inactiveMaterials.push(materialsObj[mat][matType][matTypeColor].name_str)
+                  inactiveMaterials.push(materialsObj[mat][matType][matTypeColor].name_str);
                 }
                 if (materialsObj[mat][matType][matTypeColor].status === 'discontinued') {
-                  discontinuedMaterials.push(materialsObj[mat][matType][matTypeColor].name_str || materialsObj[mat][matType][matTypeColor].material)
+                  discontinuedMaterials.push(materialsObj[mat][matType][matTypeColor].name_str || materialsObj[mat][matType][matTypeColor].material);
                 }
               }
             }
@@ -176,7 +177,9 @@ export class Feature {
   }
 
   checkMaterialsUsed() {
-    if (this.feature_type === 'seeyond') { return; }
+    if (this.feature_type === 'seeyond') {
+      return;
+    }
     let alertStr;
     const gridData = this.gridData;
     const matchedInactiveMaterials = [];
@@ -185,88 +188,100 @@ export class Feature {
     if (this.inactiveMaterials.length > 0) {
       // loop through gridData looking for inactive materials
       this.inactiveMaterials.map(material => {
-        const mat = material.toString().toLowerCase().replace(/ /g, '_');
+        const mat = material
+          .toString()
+          .toLowerCase()
+          .replace(/ /g, '_');
         gridData.map(gridSection => {
           let gridSectionArr = [];
           if (!Array.isArray(gridSection)) {
             gridSectionArr = Object.keys(gridSection).map(key => {
               return gridSection[key];
-            })
-          } else { gridSectionArr = gridSection; }
+            });
+          } else {
+            gridSectionArr = gridSection;
+          }
           gridSectionArr.map(tile => {
             if (tile.material === mat) {
               if (matchedInactiveMaterials.indexOf(material) < 0) {
                 matchedInactiveMaterials.push(material);
               }
             }
-          })
-        })
-      })
+          });
+        });
+      });
       // alert users if inactive materials are being used
       if (matchedInactiveMaterials.length === 1) {
-        this.alert.error(`${matchedInactiveMaterials[0]} is being discontinued and is only available while supplies last.`)
+        this.alert.error(`${matchedInactiveMaterials[0]} is being discontinued and is only available while supplies last.`);
       } else if (matchedInactiveMaterials.length > 1) {
         alertStr = matchedInactiveMaterials.toString();
         alertStr = alertStr.replace(/,/g, ' and ');
-        this.alert.error(`${alertStr} are being discontinued and are only available while supplies last.`)
+        this.alert.error(`${alertStr} are being discontinued and are only available while supplies last.`);
       }
     }
     if (this.discontinuedMaterials.length > 0) {
       // loop through gridData looking for discontinued materials
       this.discontinuedMaterials.map(material => {
-        const mat = material.toString().toLowerCase().replace(/ /g, '_');
+        const mat = material
+          .toString()
+          .toLowerCase()
+          .replace(/ /g, '_');
         gridData.map(gridSection => {
           let gridSectionArr = [];
           if (!Array.isArray(gridSection)) {
             gridSectionArr = Object.keys(gridSection).map(key => {
               return gridSection[key];
-            })
-          } else { gridSectionArr = gridSection; }
+            });
+          } else {
+            gridSectionArr = gridSection;
+          }
           gridSectionArr.map(tile => {
             if (tile.material === mat) {
               if (matchedDiscontinuedMaterials.indexOf(material) < 0) {
                 matchedDiscontinuedMaterials.push(material);
               }
             }
-          })
-        })
-      })
+          });
+        });
+      });
       // if discontinued materials are found disable quote and alert user
       if (matchedDiscontinuedMaterials.length > 0) {
         this.canQuote = false;
         if (matchedDiscontinuedMaterials.length === 1) {
-          this.alert.error(`The ${matchedDiscontinuedMaterials[0]} material has been discontinued. Select a new color to proceed.`)
+          this.alert.error(`The ${matchedDiscontinuedMaterials[0]} material has been discontinued. Select a new color to proceed.`);
         } else if (matchedDiscontinuedMaterials.length > 1) {
           alertStr = matchedDiscontinuedMaterials.toString();
           alertStr = alertStr.replace(/,/g, ' and ');
-          this.alert.error(`The ${alertStr} materials have been discontinued. Select a new color to proceed.`)
+          this.alert.error(`The ${alertStr} materials have been discontinued. Select a new color to proceed.`);
         }
       }
-    } else { // canQuote if no discontinuedMaterials found
+    } else {
+      // canQuote if no discontinuedMaterials found
       this.canQuote = true;
     }
   }
 
   getTetriaEstimate(tilesArray) {
-    const flatTilePrice = 61.80;
-    const tetriaTilePrice = 82.40;
+    const flatTilePrice = 61.8;
+    const tetriaTilePrice = 82.4;
     let flatTileCount = 0;
     let tetriaTileCount = 0;
     const tetriaTiles = ['01', '02', '03'];
 
     for (const tile in tilesArray) {
       if (tilesArray.hasOwnProperty(tile)) {
-        const currentTile = tilesArray[tile];
-        if (tetriaTiles.indexOf(currentTile.tile) !== -1) {
+        const currentTile = typeof tilesArray[tile].tile === 'string' ? tilesArray[tile].tile : tilesArray[tile].tile.tile_size;
+        const purchased = typeof tilesArray[tile].tile === 'string' ? tilesArray[tile].purchased : tilesArray[tile].purchased;
+        if (tetriaTiles.indexOf(currentTile) !== -1) {
           // add the purchased amount to the tetria tile count
-          tetriaTileCount += currentTile.purchased;
-        }else if (currentTile.tile === '00') {
+          tetriaTileCount += purchased;
+        } else if (currentTile === '00') {
           // add the purchased amount to the flat tile count
-          flatTileCount += currentTile.purchased;
+          flatTileCount += purchased;
         }
       }
     }
-    this.services_amount = (tetriaTileCount * tetriaTilePrice) + (flatTileCount * flatTilePrice);
+    this.services_amount = tetriaTileCount * tetriaTilePrice + flatTileCount * flatTilePrice;
     this.estimated_amount = this.services_amount;
   }
 
@@ -280,7 +295,7 @@ export class Feature {
     }
 
     const hardware110 = 0.23;
-    const hardware111 = 1.80;
+    const hardware111 = 1.8;
     const hardware112 = 3.08;
     const total110 = hardware110 * hushTileCount * 5;
     const total111 = hardware111 * hushTileCount * 4;
@@ -290,30 +305,30 @@ export class Feature {
       '3-85-110': hushTileCount * 5,
       '3-85-111': hushTileCount * 4,
       '3-85-112': hushTileCount * 2
-    }
+    };
 
     const allHardwareCost = total110 + total111 + total112;
-    this.services_amount = (hushTileCount * 65.49);
+    this.services_amount = hushTileCount * 65.49;
     this.estimated_amount = this.services_amount + allHardwareCost;
   }
 
   getClarioEstimate(tilesArray) {
-    let products_amount = 0.00;
+    let products_amount = 0.0;
     let clario24TileCount = 0;
     let clario48TileCount = 0;
     let clario00TileCount = 0;
     let sheetsNeeded = 0;
-    let sheetCost = 0.00;
+    let sheetCost = 0.0;
     for (const tile in tilesArray) {
       if (tilesArray.hasOwnProperty(tile)) {
         const currentTile = tilesArray[tile];
-        if (currentTile.tile === '24') {
+        if (currentTile.tile === '24' || currentTile.tile === '600' || currentTile.tile === '625') {
           // 24x24 prices
           clario24TileCount += currentTile.purchased;
           // what part_id is the material?
           // how many sheets do we need? sheetsNeeded = (currentTile.purchased / 4);
           sheetsNeeded = currentTile.purchased / 4;
-        }else if (currentTile.tile === '48') {
+        } else if (currentTile.tile === '48' || currentTile.tile === '1200' || currentTile.tile === '1250') {
           // 24x48 prices
           clario48TileCount += currentTile.purchased;
           sheetsNeeded = currentTile.purchased / 2;
@@ -352,7 +367,7 @@ export class Feature {
     let variaSheetsNeeded: number;
     let variaDiffusionSheetsNeeded: number;
     const variaSheetCost = 488.14;
-    const variaDiffusionSheetCost: number = variaSheetCost + 100.00;
+    const variaDiffusionSheetCost: number = variaSheetCost + 100.0;
 
     for (const tile in tilesArray) {
       if (tilesArray.hasOwnProperty(tile)) {
@@ -373,23 +388,23 @@ export class Feature {
     variaDiffusionSheetsNeeded = Math.ceil(veloVariaDiffusionTiles / 8);
     this.debug.log('feature', `varia sheets needed, ${variaSheetsNeeded}`);
     this.debug.log('feature', `varia diffusion sheets needed ${variaDiffusionSheetsNeeded}`);
-    products_amount = (variaSheetsNeeded * variaSheetCost) + (variaDiffusionSheetsNeeded * variaDiffusionSheetCost);
+    products_amount = variaSheetsNeeded * variaSheetCost + variaDiffusionSheetsNeeded * variaDiffusionSheetCost;
 
     // SERVICES AMOUNT
     const veloFeltServiceCost = 77.25;
     const veloVariaServiceCost = 78.75;
-    this.services_amount = (veloFeltTiles * veloFeltServiceCost) + ((veloVariaTiles + veloVariaDiffusionTiles) * veloVariaServiceCost);
+    this.services_amount = veloFeltTiles * veloFeltServiceCost + (veloVariaTiles + veloVariaDiffusionTiles) * veloVariaServiceCost;
     // this.debug.logfeature', ('=== SERVICES AMOUNT ===');
     // this.debug.logfeature', (this.services_amount);
 
     // HARDWARE AMOUNT
     let hardware_amount: number;
-    let hardwareCost = 0.00;
+    let hardwareCost = 0.0;
     let cableCount: number;
-    let cableCost = 0.00;
+    let cableCost = 0.0;
     const cableKitCost = 12.46;
     const variaConnectionKitCost = 6.85;
-    const feltConnectionKitCost = .46;
+    const feltConnectionKitCost = 0.46;
     const drillBitCost = 10.23;
     const variaPunchToolCost = 17.49;
     let variaConnectionKitsNeeded = 0;
@@ -414,8 +429,8 @@ export class Feature {
         // if ratio > 1 then cableCount = Math.ceil(cables * .5)
         // this is the total number of purchased tiles
         // this is the number of tiles in the design
-        const ratio = (sharedEdges) / (tilesInIsland);
-        const factor = ratio < 1 ? .75 : .5;
+        const ratio = sharedEdges / tilesInIsland;
+        const factor = ratio < 1 ? 0.75 : 0.5;
         cableCount = Math.ceil(tilesInIsland * factor);
 
         // If shared edges is 1 less than total tiles, set cableCount to sharedEdges
@@ -430,7 +445,9 @@ export class Feature {
         cablesNeeded += cableCount;
 
         // Calculate the hardware cost for connections and add to the hardware cost
-        hardwareCost += (islandConnections['variaToVaria'] * variaConnectionKitCost) + ((islandConnections['feltToFelt'] + islandConnections['variaToFelt']) * feltConnectionKitCost);
+        hardwareCost +=
+          islandConnections['variaToVaria'] * variaConnectionKitCost +
+          (islandConnections['feltToFelt'] + islandConnections['variaToFelt']) * feltConnectionKitCost;
 
         // Add the connections to the running total
         variaConnectionKitsNeeded += islandConnections['variaToVaria'];
@@ -466,14 +483,14 @@ export class Feature {
       '3-15-8899-K': variaConnectionKitsNeeded,
       '3-85-105-K': feltConnectionKitsNeeded,
       '3-15-8813': variaPunchToolNeeded ? 1 : 0
-    }
+    };
 
     this.debug.log('feature', '=====feature HARDWARE =====');
     this.debug.log('feature', `${this.hardware}`);
     this.debug.log('feature', '=====feature END HARDWARE =====');
   }
 
-  updateSelectedTile(tile: string) {
+  updateSelectedTile(tile) {
     this.selectedTile = tile;
 
     // if a tool is selected then remove it
@@ -501,7 +518,9 @@ export class Feature {
     // handle no color for varia
     if (material === 'no_color' && materialType === 'varia') {
       // forces a diffusion to be selected if 'no_color' is chosen
-      if (!this.diffusion) { this.updateSelectedDiffusion('avalanche_d01'); }
+      if (!this.diffusion) {
+        this.updateSelectedDiffusion('avalanche_d01');
+      }
     }
   }
 
@@ -522,7 +541,7 @@ export class Feature {
       // a diffusion requires varia to be selected
       this.updateSelectedMaterial('no_color', '#ffffff', 'varia');
     }
-    const hasColor = (this.material !== 'no_color') ? true : false;
+    const hasColor = this.material !== 'no_color' ? true : false;
     // if the diffusion they clicked on is already selected,
     // deselect it so they have a way to remove the diffusion
     // unless 'no_color' is selected
@@ -545,12 +564,13 @@ export class Feature {
 
   clearAll() {
     this.gridData = undefined;
-    this.estimated_amount = 0.00;
+    this.estimated_amount = 0.0;
     this.buildGrid();
   }
 
   applyAll() {
     this.updateEstimatedAmount();
+    this.clearAll();
     this.onApplyAll.emit();
   }
 
@@ -570,31 +590,56 @@ export class Feature {
   }
 
   public getRows() {
-    let rows: number;
-
-    // velo has a static grid
-    if (this.feature_type === 'velo') {
-      rows = 500;
-    }else if (this.units === 'inches') {
-      rows = Math.ceil(this.length / 12 / 2);
-    } else {
-      rows = Math.ceil(this.convertCMtoIN(this.length) / 12 / 2);
+    if (!!this.length) {
+      let rows: number;
+      if (this.feature_type === 'clario') {
+        return this.getClarioGridSize('row');
+      }
+      // velo has a static grid
+      if (this.feature_type === 'velo') {
+        rows = 500;
+      } else if (this.units === 'inches') {
+        rows = Math.ceil(this.length / 12 / 2);
+      } else {
+        rows = Math.ceil(this.convertCMtoIN(this.length) / 12 / 2);
+      }
+      return rows;
     }
-    return rows;
   }
 
   public getColumns() {
-    let columns: number;
-
-    // velo has a static grid
-    if (this.feature_type === 'velo') {
-      columns = 820;
-    }else if (this.units === 'inches') {
-      columns = Math.ceil(this.width / 12 / 2);
-    } else {
-      columns = Math.ceil(this.convertCMtoIN(this.width) / 12 / 2);
+    if (!!this.width) {
+      let columns: number;
+      if (this.feature_type === 'clario') {
+        return this.getClarioGridSize('column');
+      }
+      // velo has a static grid
+      if (this.feature_type === 'velo') {
+        columns = 820;
+      } else if (this.units === 'inches') {
+        columns = Math.ceil(this.width / 12 / 2);
+      } else {
+        columns = Math.ceil(this.convertCMtoIN(this.width) / 12 / 2);
+      }
+      return columns;
     }
-    return columns;
+  }
+
+  public getClarioGridSize(dimension) {
+    const size = dimension === 'row' ? this.length : this.width;
+    let count;
+    switch (this.clairoTileSizeType) {
+      case 'standard':
+        count = Math.ceil(size / 12 / 2);
+        break;
+      case 'metric':
+        count = Math.ceil((size * 10) / 600);
+        break;
+      case 'german':
+        count = Math.ceil((size * 10) / 625);
+        break;
+    }
+    return count;
   }
 
   public getFeatureTypeInteger() {
@@ -637,17 +682,23 @@ export class Feature {
 
   public getPackageQty(tile: string) {
     let qty: number;
-    if (this.feature_type === 'hush') { return 1; }
+    if (this.feature_type === 'hush') {
+      return 1;
+    }
     switch (tile) {
       case '00':
       case '01':
       case '02':
       case '03':
       case '24':
+      case '600':
+      case '625':
         qty = 4;
         break;
 
       case '48':
+      case '1200':
+      case '1250':
         qty = 2;
         break;
 
@@ -682,7 +733,9 @@ export class Feature {
 
   public getTilesPurchasedObj() {
     let tiles: {};
-    if (this.is_quantity_order) { return; }
+    if (this.is_quantity_order) {
+      return;
+    }
     if (this.feature_type === 'velo') {
       const pkgQty: number = this.getPackageQty('velo');
       const gridTiles = this.veloTiles();
@@ -693,54 +746,119 @@ export class Feature {
           const materialType = gridTiles[tile].materialType;
           const material = gridTiles[tile].material;
           const diffusion = gridTiles[tile].diffusion || '';
-          const key = (!diffusion) ? `${materialType}-${material}` : `${materialType}-${material}-${diffusion}`;
-          if (purchasedTiles === undefined) { purchasedTiles = {}; }
+          const key = !diffusion ? `${materialType}-${material}` : `${materialType}-${material}-${diffusion}`;
+          if (purchasedTiles === undefined) {
+            purchasedTiles = {};
+          }
           if (!!purchasedTiles[key]) {
             purchasedTiles[key][gridTiles[tile].tile] += 1;
             purchasedTiles[key].purchased = pkgQty * Math.ceil((purchasedTiles[key].concave + purchasedTiles[key].convex) / pkgQty);
           } else {
             purchasedTiles[key] = {
-              'purchased': pkgQty,
-              'image': gridTiles[tile].materialType === 'felt' ? '/assets/images/materials/felt/merino/' + gridTiles[tile].material + '.png' : '/assets/images/tiles/00/' + gridTiles[tile].material + '.png',
-              'hex': gridTiles[tile].materialType === 'varia' ? gridTiles[tile].hex : '',
-              'convex': gridTiles[tile].tile === 'convex' ? 1 : 0,
-              'concave': gridTiles[tile].tile === 'concave' ? 1 : 0,
-              'material': gridTiles[tile].material,
-              'materialType': gridTiles[tile].materialType,
-              'tile': gridTiles[tile].tile,
-              'diffusion': gridTiles[tile].diffusion
-            }
+              purchased: pkgQty,
+              image:
+                gridTiles[tile].materialType === 'felt'
+                  ? '/assets/images/materials/felt/merino/' + gridTiles[tile].material + '.png'
+                  : '/assets/images/tiles/00/' + gridTiles[tile].material + '.png',
+              hex: gridTiles[tile].materialType === 'varia' ? gridTiles[tile].hex : '',
+              convex: gridTiles[tile].tile === 'convex' ? 1 : 0,
+              concave: gridTiles[tile].tile === 'concave' ? 1 : 0,
+              material: gridTiles[tile].material,
+              materialType: gridTiles[tile].materialType,
+              tile: gridTiles[tile].tile,
+              diffusion: gridTiles[tile].diffusion
+            };
           }
         }
       }
       tiles = purchasedTiles;
+    } else if (this.feature_type === 'clario') {
+      const filteredGridData = [];
+      // populate filteredGridData with empty GridSections
+      for (let r = 0; r < this.getRows(); r++) {
+        filteredGridData[r] = [];
+        for (let c = 0; c < this.getColumns(); c++) {
+          filteredGridData[r][c] = new GridSection(r, c);
+        }
+      }
+      const gridIds = [];
+      this.gridData.map(r => {
+        r.map(c => {
+          if (c.gridTileID) {
+            // use the gridTileID if it has one
+            if (c.gridTileID === 0 || gridIds.includes(c.gridTileID)) {
+              return;
+            }
+            gridIds.push(c.gridTileID);
+            filteredGridData[c.row][c.column] = this.gridData[c.row][c.column];
+          } else if (c.tile) {
+            // if there is no gridTileID make assumptions based off the backgroundImage
+            if (c.tileSize === '48') {
+              // if it's a 48 tile remove it's companion
+              if (!!filteredGridData[c.row][c.column + 1] && filteredGridData[c.row][c.column + 1].backgroundImage === c.backgroundImage) {
+                filteredGridData[c.row][c.column + 1] = new GridSection(c.row, c.column + 1);
+              } else if (!!filteredGridData[c.row][c.column - 1] && filteredGridData[c.row][c.column - 1].backgroundImage === c.backgroundImage) {
+                filteredGridData[c.row][c.column - 1] = new GridSection(c.row, c.column - 1);
+              }
+            }
+            filteredGridData[c.row][c.column] = this.gridData[c.row][c.column];
+          }
+        });
+      });
+      // Determine the number of unique tiles (color and tile)
+      let pkgQty: number;
+      if (filteredGridData) {
+        for (let i = filteredGridData.length - 1; i >= 0; i--) {
+          for (let j = filteredGridData[i].length - 1; j >= 0; j--) {
+            if (filteredGridData[i][j].tile) {
+              const key = filteredGridData[i][j]['material'] + '-' + filteredGridData[i][j]['tile'];
+              pkgQty = this.getPackageQty(filteredGridData[i][j]['tile']);
+              if (tiles === undefined) {
+                tiles = {};
+              }
+              if (!!tiles[key]) {
+                tiles[key].used += 1;
+                tiles[key].purchased = pkgQty * Math.ceil(tiles[key].used / pkgQty);
+              } else {
+                const tileType = filteredGridData[i][j]['tile'] === '00' ? 'tiles' : this.getTileType('plural');
+                const imageUrl = this.getClarioImgUrl(filteredGridData[i][j].tile, filteredGridData[i][j]['material']);
+                tiles[key] = {
+                  purchased: pkgQty,
+                  image: imageUrl,
+                  used: 1,
+                  material: filteredGridData[i][j]['material'],
+                  tile: filteredGridData[i][j]['tile']
+                };
+              }
+            }
+          }
+        }
+      }
     } else {
       // Determine the number of unique tiles (color and tile)
       let pkgQty: number;
-      let tileType = this.getTileType('plural');
       if (this.gridData) {
         for (let i = this.gridData.length - 1; i >= 0; i--) {
           for (let j = this.gridData[i].length - 1; j >= 0; j--) {
             if (this.gridData[i][j].tile) {
               const key = this.gridData[i][j]['material'] + '-' + this.gridData[i][j]['tile'];
-                pkgQty = this.getPackageQty(this.gridData[i][j]['tile']);
-              if (tiles === undefined) { tiles = {}; }
+              pkgQty = this.getPackageQty(this.gridData[i][j]['tile']);
+              if (tiles === undefined) {
+                tiles = {};
+              }
               if (!!tiles[key]) {
                 tiles[key].used += 1;
                 tiles[key].purchased = pkgQty * Math.ceil(tiles[key].used / pkgQty);
               } else {
-                if (this.gridData[i][j]['tile'] === '00') {
-                  tileType = 'tiles';
-                } else {
-                  tileType = this.getTileType('plural');
-                }
+                const tileType = this.gridData[i][j]['tile'] === '00' ? 'tiles' : this.getTileType('plural');
+                const imageUrl = `/assets/images/${tileType}/${this.gridData[i][j].tile}/${this.gridData[i][j]['material']}.png`;
                 tiles[key] = {
-                  'purchased': pkgQty,
-                  'image': '/assets/images/' + tileType + '/' + this.gridData[i][j]['tile'] + '/' + this.gridData[i][j]['material'] + '.png',
-                  'used': 1,
-                  'material': this.gridData[i][j]['material'],
-                  'tile': this.gridData[i][j]['tile']
-                }
+                  purchased: pkgQty,
+                  image: imageUrl,
+                  used: 1,
+                  material: this.gridData[i][j]['material'],
+                  tile: this.gridData[i][j]['tile']
+                };
               }
             }
           }
@@ -749,6 +867,21 @@ export class Feature {
     }
     this.tiles = tiles;
     return tiles;
+  }
+
+  public getClarioImgUrl(tileSize, material) {
+    let tileType;
+    let tileImageType;
+    if (tileSize === '00') {
+      // flat tile selected
+      tileType = 'tiles';
+      tileImageType = '00';
+    } else {
+      tileType = 'baffles';
+      const squareImgs = ['24', '600', '625'];
+      tileImageType = squareImgs.includes(tileSize) ? '24' : '48';
+    }
+    return `/assets/images/${tileType}/${tileImageType}/${material}.png`;
   }
 
   public getPurchasedVeloTiles(materialType: string) {
@@ -765,19 +898,28 @@ export class Feature {
   public getUserInputs() {
     let tiles;
     switch (this.feature_type) {
-      case 'velo': tiles = this.veloTiles(); break;
-      case 'hush': tiles = this.hushTiles(); break;
-      default: tiles = this.gridData; break;
+      case 'velo':
+        tiles = this.veloTiles();
+        break;
+      case 'hush':
+        tiles = this.hushTiles();
+        break;
+      case 'clario':
+        tiles = this.clarioTiles();
+        break;
+      default:
+        tiles = this.gridData;
+        break;
     }
 
     return {
-      'UserInputs': {
-        'Type': this.getFeatureTypeInteger(),
-        'NumX': this.getColumns(),
-        'NumY': this.getRows(),
-        'Tiles': tiles
+      UserInputs: {
+        Type: this.getFeatureTypeInteger(),
+        NumX: this.getColumns(),
+        NumY: this.getRows(),
+        Tiles: tiles
       }
-    }
+    };
   }
 
   public convertCMtoIN(cm: number) {
@@ -796,7 +938,7 @@ export class Feature {
 
   public veloTiles() {
     const veloTiles = [];
-    for ( const tile in this.gridData) {
+    for (const tile in this.gridData) {
       if (this.gridData[tile].texture !== '') {
         veloTiles.push(this.gridData[tile]);
       }
@@ -816,6 +958,32 @@ export class Feature {
     return hushTiles;
   }
 
+  public clarioTiles() {
+    const clarioTiles = this.gridData.slice(0);
+    const tileIds = [];
+    for (let i = 0; i < clarioTiles.length; i++) {
+      for (let j = 0; j < clarioTiles[i].length; j++) {
+        if (clarioTiles[i][j].tile !== '') {
+          const squareImgs = ['24', '600', '625'];
+          const rectImgs = ['48', '1200', '1250'];
+          if (squareImgs.includes(clarioTiles[i][j].tile)) {
+            clarioTiles[i][j].tile = '24';
+          } else if (rectImgs.includes(clarioTiles[i][j].tile)) {
+            clarioTiles[i][j].tile = '48';
+          } else {
+            clarioTiles[i][j].tile = '00';
+          }
+        }
+        if (tileIds.includes(clarioTiles[i][j].gridTileID)) {
+          clarioTiles[i][j].texture = '';
+        } else {
+          tileIds.push(clarioTiles[i][j].gridTileID);
+        }
+      }
+    }
+    return clarioTiles;
+  }
+
   public findVeloTileAt(x, y) {
     for (const el in this.gridData) {
       if (this.gridData[el].x === x && this.gridData[el].y === y) {
@@ -824,7 +992,7 @@ export class Feature {
     }
   }
 
-  public getVeloConnections(island: any): any [] {
+  public getVeloConnections(island: any): any[] {
     const veloTiles = [];
     let veloConnections: any;
     let variaToVariaCount = 0;
@@ -850,7 +1018,7 @@ export class Feature {
               const neighborIndex = neighbor.index;
               const a = Math.min(thisIndex, neighborIndex);
               const b = Math.max(thisIndex, neighborIndex);
-              const mappedIndex = (a + b) * (a + b + 1) / 2 + a;
+              const mappedIndex = ((a + b) * (a + b + 1)) / 2 + a;
               if (typeof neighbor.materialType !== 'undefined' && !matches[mappedIndex]) {
                 // felt to felt seams
                 if (thisMaterialType === 'felt' && neighbor.materialType === 'felt') {
@@ -878,10 +1046,10 @@ export class Feature {
     }
 
     veloConnections = {
-      'variaToVaria': variaToVariaCount,
-      'variaToFelt': variaToFeltCount,
-      'feltToFelt': feltToFeltCount,
-      'totalConnections': variaToVariaCount + variaToFeltCount + feltToFeltCount
+      variaToVaria: variaToVariaCount,
+      variaToFelt: variaToFeltCount,
+      feltToFelt: feltToFeltCount,
+      totalConnections: variaToVariaCount + variaToFeltCount + feltToFeltCount
     };
     return veloConnections;
   }
@@ -894,7 +1062,9 @@ export class Feature {
       const index = indices[i];
       const island = this._getIsland(+index);
 
-      if (island.length <= 0) { continue };
+      if (island.length <= 0) {
+        continue;
+      }
 
       indices = _.difference(indices, island);
       islands.push(island);
@@ -902,7 +1072,7 @@ export class Feature {
     return islands;
   }
 
-  private _getIsland(index: number, members: any = []): any [] {
+  private _getIsland(index: number, members: any = []): any[] {
     const tileObject = this.gridData[index];
     if (tileObject.texture === '') {
       return members;
@@ -983,11 +1153,11 @@ export class Feature {
       info = 'Tiles are sold in quanties of 4.';
     }
 
-    if (this.feature_type === 'clario' && this.tile_size === 24) {
+    if (this.feature_type === 'clario' && this.tile_image_type === 24) {
       info = 'Baffles are sold in quantities of 4.';
     }
 
-    if (this.feature_type === 'clario' && this.tile_size === 48) {
+    if (this.feature_type === 'clario' && this.tile_image_type === 48) {
       info = '24x24 baffles are sold in qty of 4, and 24x48 baffles are sold in qty of 2.';
     }
 
@@ -999,12 +1169,12 @@ export class Feature {
   }
 
   public updateGridUnits(units: string) {
-    if ( this.feature_type === 'velo' ) {
-      if (units === 'centimeters' && this.units !== units ) {
+    if (this.feature_type === 'velo') {
+      if (units === 'centimeters' && this.units !== units) {
         // convert measurements to cm
         this.width = 976;
         this.length = 610;
-      }else if (units === 'inches' && this.units !== units) {
+      } else if (units === 'inches' && this.units !== units) {
         // convert measurement to inches
         this.width = 384;
         this.length = 240;
@@ -1015,7 +1185,9 @@ export class Feature {
   }
 
   public setFeatureType(str: string) {
-    if (str.indexOf('hush') > -1) { str = 'hush'; }
+    if (str.indexOf('hush') > -1) {
+      str = 'hush';
+    }
     this.feature_type = str;
     return str;
   }
@@ -1030,23 +1202,37 @@ export class Feature {
     const featureType = this.feature_type;
     let requiredMaterials: any;
     switch (featureType) {
-      case 'hush': requiredMaterials = this.materials.felt.sola; break;
-      case 'seeyond': requiredMaterials = this.materials.felt.sola; break;
-      case 'tetria': requiredMaterials = this.materials.felt.merino; break;
-      case 'clario': requiredMaterials = this.materials.felt.sola; break;
+      case 'hush':
+        requiredMaterials = this.materials.felt.sola;
+        break;
+      case 'seeyond':
+        requiredMaterials = this.materials.felt.sola;
+        break;
+      case 'tetria':
+        requiredMaterials = this.materials.felt.merino;
+        break;
+      case 'clario':
+        requiredMaterials = this.materials.felt.sola;
+        break;
       case 'velo':
-        requiredMaterials = {felt: undefined, varia: undefined};
+        requiredMaterials = { felt: undefined, varia: undefined };
         requiredMaterials.felt = this.materials.felt.merino;
         this.materials.varia.color = this.addNoColorToVariaObj();
         requiredMaterials.varia = this.materials.varia;
-      break;
+        break;
     }
     return requiredMaterials;
   }
 
   addNoColorToVariaObj() {
     // object to add
-    const variaWithoutColor = { material: 'no_color', hex: '#ffffff', status: 'active', availableUntil: '', partId: null };
+    const variaWithoutColor = {
+      material: 'no_color',
+      hex: '#ffffff',
+      status: 'active',
+      availableUntil: '',
+      partId: null
+    };
     // turn it into an array to enforce object order
     const variaArr = Object.keys(this.materials.varia.color).map(key => this.materials.varia.color[key]);
     // add object the end of the array
@@ -1058,5 +1244,10 @@ export class Feature {
     }, {});
     // return the new object with 'no_color' added to the end of the array
     return newVariaObj;
+  }
+
+  navToLanding() {
+    this.location.go('/');
+    window.location.reload();
   }
 }
