@@ -22,8 +22,8 @@ export class LoginComponent implements OnInit {
     private debug: DebugService,
     public feature: Feature,
     public user: User,
-    public dialogRef: MatDialogRef<LoginComponent>,
-  ) { }
+    public dialogRef: MatDialogRef<LoginComponent>
+  ) {}
 
   ngOnInit() {
     this.debug.log('login-component', this.user);
@@ -31,29 +31,26 @@ export class LoginComponent implements OnInit {
 
   login() {
     this.loading = true;
-    this.api.login(this.email, this.password)
-      .subscribe(
-        data => {
-          if (!!data.result && !!data.result.user) {
-            this.alert.success('Successfully logged in.');
-            this.loading = false;
-            localStorage.setItem('3formUser', JSON.stringify(data.result.user));
-            this.api.onUserLoggedIn.emit(this.user);
-            this.debug.log('api', 'user successfully logged in');
-            this.dialogRef.close();
-          } else {
-            this.alert.error('Incorrect Username or Password');
-          }
+    this.api.login(this.email, this.password).subscribe(
+      data => {
+        if (!!data.result && !!data.result.user) {
+          this.alert.success('Successfully logged in.');
           this.loading = false;
-        },
-        error => {
-          if (error) {
-            this.api.handleError(error);
-          }
-          this.loading = false;
+          localStorage.setItem('3formUser', JSON.stringify(data.result.user));
+          this.api.onUserLoggedIn.emit(this.user);
+          this.debug.log('api', 'user successfully logged in');
+          this.dialogRef.close();
+        } else {
+          this.alert.error('Incorrect Username or Password');
         }
-      );
+        this.loading = false;
+      },
+      error => {
+        if (error) {
+          this.api.handleError(error);
+        }
+        this.loading = false;
+      }
+    );
   }
-
 }
-
