@@ -1,3 +1,4 @@
+import { ClarioGridsService } from './../../_services/clario-grids.service';
 import { DebugService } from './../../_services/debug.service';
 import { Feature } from './../../feature';
 import { Component, OnInit, Inject, AfterContentInit } from '@angular/core';
@@ -19,7 +20,7 @@ export class AddQuantityComponent implements OnInit, AfterContentInit {
   selectedMaterial: string;
   selectedMaterialImg: string;
   selectedQuantity: number;
-  selectedTile = '00';
+  // selectedTile: any;
 
   constructor(
     @Inject(MAT_DIALOG_DATA) public inputRow: any,
@@ -27,8 +28,9 @@ export class AddQuantityComponent implements OnInit, AfterContentInit {
     public materialsService: MaterialsService,
     public qtySrv: QuantityService,
     public feature: Feature,
-    public debug: DebugService
-  ) { }
+    public debug: DebugService,
+    public clarioGrids: ClarioGridsService
+  ) {}
 
   ngOnInit() {
     this.getFeatureMaterials();
@@ -36,13 +38,18 @@ export class AddQuantityComponent implements OnInit, AfterContentInit {
   }
 
   ngAfterContentInit() {
-    if (!!this.inputRow) { this.loadRowForEdit(); }
+    if (!!this.inputRow) {
+      setTimeout(() => {
+        this.loadRowForEdit();
+      });
+    }
   }
 
   loadRowForEdit() {
     this.debug.log('quantity', this.inputRow);
     this.isEditing = true;
     this.updateSelectedMaterial(this.inputRow.material);
+    this.updateSelectedTile(this.feature.selectedTile);
     this.quantityDidChange(this.inputRow.used);
     this.updateSelectedTile(this.inputRow.tile);
   }
@@ -50,9 +57,15 @@ export class AddQuantityComponent implements OnInit, AfterContentInit {
   getFeatureMaterials() {
     let requiredMaterials: any;
     switch (this.qtySrv.feature_type) {
-      case 'hush': requiredMaterials = this.materialsService.materials.felt.sola; break;
-      case 'tetria': requiredMaterials = this.materialsService.materials.felt.merino; break;
-      case 'clario': requiredMaterials = this.materialsService.materials.felt.sola; break;
+      case 'hush':
+        requiredMaterials = this.materialsService.materials.felt.sola;
+        break;
+      case 'tetria':
+        requiredMaterials = this.materialsService.materials.felt.merino;
+        break;
+      case 'clario':
+        requiredMaterials = this.materialsService.materials.felt.sola;
+        break;
     }
     this.materials = requiredMaterials;
   }
@@ -64,17 +77,33 @@ export class AddQuantityComponent implements OnInit, AfterContentInit {
 
   updateMaterialImg() {
     let materialImg;
-    const tileType = (this.selectedTile === '24' || this.selectedTile === '48') ? 'baffles' : 'tiles';
     switch (this.qtySrv.feature_type) {
-      case 'hush': materialImg = `/assets/images/${tileType}/${this.selectedTile}/${this.selectedMaterial}.png`; break;
-      case 'tetria': materialImg = `/assets/images/${tileType}/${this.selectedTile}/${this.selectedMaterial}.png`; break; // TODO FIX THIS
-      case 'clario': materialImg = `/assets/images/${tileType}/${this.selectedTile}/${this.selectedMaterial}.png`; break; // TODO FIX THIS
+      case 'hush':
+        materialImg = `/assets/images/tiles/${this.feature.selectedTile.tile}/${this.selectedMaterial}.png`;
+        break;
+      case 'tetria':
+        materialImg = `/assets/images/tiles/${this.feature.selectedTile.tile}/${this.selectedMaterial}.png`;
+        break;
+      case 'clario':
+        let tileType;
+        let tileImageType;
+        if (this.feature.selectedTile.tile_size === '00') {
+          // flat tile selected
+          tileType = 'tiles';
+          tileImageType = '00';
+        } else {
+          tileType = 'baffles';
+          const squareImgs = ['24', '600', '625'];
+          tileImageType = squareImgs.includes(this.feature.selectedTile.tile_size) ? '24' : '48';
+        }
+        materialImg = `/assets/images/${tileType}/${tileImageType}/${this.selectedMaterial}.png`;
+        break;
     }
     this.selectedMaterialImg = materialImg;
   }
 
   updateSelectedTile(tile) {
-    this.selectedTile = tile;
+    this.feature.updateSelectedTile(tile);
     this.updateMaterialImg();
   }
 
@@ -85,9 +114,15 @@ export class AddQuantityComponent implements OnInit, AfterContentInit {
   validateQtyInputs() {
     let isValid = false;
     switch (this.qtySrv.feature_type) {
-      case 'hush': isValid = (!!this.selectedMaterial && (this.selectedQuantity > 0)); break;
-      case 'tetria': isValid = (!!this.selectedMaterial && (this.selectedQuantity > 0)); break; // TODO FIX THIS
-      case 'clario': isValid = (!!this.selectedMaterial && (this.selectedQuantity > 0)); break; // TODO FIX THIS
+      case 'hush':
+        isValid = !!this.selectedMaterial && this.selectedQuantity > 0;
+        break;
+      case 'tetria':
+        isValid = !!this.selectedMaterial && this.selectedQuantity > 0;
+        break; // TODO FIX THIS
+      case 'clario':
+        isValid = !!this.selectedMaterial && this.selectedQuantity > 0;
+        break; // TODO FIX THIS
     }
     return isValid;
   }
@@ -99,15 +134,19 @@ export class AddQuantityComponent implements OnInit, AfterContentInit {
   }
 
   addToOrder() {
-    const pkgQty = this.feature.getPackageQty(this.selectedTile);
-    const key = `${this.selectedMaterial}-${this.selectedTile}`
-    const selections = {[key]: {
-      purchased: pkgQty * Math.ceil(this.selectedQuantity / pkgQty),
-      image: this.selectedMaterialImg,
-      used: this.selectedQuantity,
-      material: this.selectedMaterial,
-      tile: this.selectedTile
-    }}
+    const pkgQty = this.feature.getPackageQty(this.feature.selectedTile.tile_size || this.feature.selectedTile.tile);
+    const key = `${this.selectedMaterial}-${this.feature.selectedTile.tile}`;
+    const tile = this.feature.feature_type === 'clario' ? this.feature.selectedTile.tile : this.feature.selectedTile;
+    const selections = {
+      [key]: {
+        purchased: pkgQty * Math.ceil(this.selectedQuantity / pkgQty),
+        image: this.selectedMaterialImg,
+        used: this.selectedQuantity,
+        material: this.selectedMaterial,
+        tile: tile,
+        tile_size: this.feature.selectedTile.tile_size || ''
+      }
+    };
     this.dialogRef.close(selections);
   }
 }
