@@ -1,3 +1,4 @@
+import { ClarioGridsService } from './../_services/clario-grids.service';
 import { TileObj } from './quantity.service';
 import { MatTableDataSource } from '@angular/material';
 import { TileRow } from './quantity.component';
@@ -12,29 +13,12 @@ export class QuantityService {
   qtyTilesArray = <TileObj[]>[];
   estimatedPrice = 0;
   tilesSelected = 0;
-  sqFtUsed = 0;
-  sqFtReceiving = 0;
-  sqFtPerTile: number;
+  sqAreaUsed = 0;
+  sqAreaReceiving = 0;
   order = new MatTableDataSource();
   rowIndexNum = 1;
 
-  constructor(
-    private debug: DebugService,
-    public feature: Feature,
-    private route: ActivatedRoute
-  ) {}
-
-  setRowData(row) {
-    this.debug.log('quantity', row);
-    this.getRowEstimate(row); // sets feature.estimated_amount
-    const newRow = row[Object.keys(row)[0]];
-    this.feature.material = newRow.material;
-    this.feature.tile_size = (newRow.tile === '48') ? 48 : 24;
-    newRow.total = this.feature.estimated_amount;
-    newRow.tileSqFt = this.getTileSqFt(newRow.tile);
-    newRow.id = this.rowIndexNum++;
-    return newRow;
-  }
+  constructor(private debug: DebugService, public feature: Feature, private route: ActivatedRoute, private clarioGrids: ClarioGridsService) {}
 
   doAddRow(row) {
     this.debug.log('quantity', row);
@@ -44,17 +28,32 @@ export class QuantityService {
     this.updateSummary();
   }
 
+  setRowData(row) {
+    this.debug.log('quantity', row);
+    this.getRowEstimate(row); // sets feature.estimated_amount
+    const newRow = row[Object.keys(row)[0]];
+    this.feature.material = newRow.material;
+    this.feature.tile_image_type = newRow.tile_image_type === '48' ? 48 : 24;
+    // newRow.tile = this.feature.feature_type === 'clario' ? this.clarioGrids.selectedTileSize.tile_size : newRow.tile_image_type;
+    newRow.total = this.feature.estimated_amount;
+    newRow.tileSqArea = this.getTileSqArea(newRow.tile);
+    newRow.id = this.rowIndexNum++;
+    newRow.material_size = typeof newRow.tile === 'string' ? newRow.tile : newRow.tile.tile;
+    return newRow;
+  }
+
   combineRows(matchedRow, requestedRow) {
     // matchedRow is kept
-    const pkgQty = this.feature.getPackageQty(matchedRow.tile);
+    const pkgQty = this.feature.getPackageQty(matchedRow.tile.tile);
     const requestedRowConfigured = this.setRowData(requestedRow);
     matchedRow.used += requestedRowConfigured.used;
     matchedRow.purchased = pkgQty * Math.ceil(matchedRow.used / pkgQty);
-    const objectKey = `${matchedRow.material}-${matchedRow.tile}`;
-    const matchedRowFmtd = {[objectKey]: matchedRow};
+    const objectKey = `${matchedRow.material}-${matchedRow.tile.tile}`;
+    const matchedRowFmtd = { [objectKey]: matchedRow };
     this.getRowEstimate(matchedRowFmtd); // sets feature.estimated_amount
     matchedRow.total = this.feature.estimated_amount;
     matchedRow.id = this.rowIndexNum++;
+    matchedRow.material_size = typeof matchedRow.tile === 'string' ? matchedRow.tile : matchedRow.tile.tile;
     this.updateSummary();
   }
 
@@ -62,18 +61,20 @@ export class QuantityService {
     const untypedData: any[] = this.order.data.slice();
     const dataImagesArr = untypedData.map(rowData => rowData.image);
     // new array of objects with number of instances of each image
-    const dataImagesCounted = dataImagesArr.reduce((a, b) => Object.assign(a, {[b]: (a[b] || 0) + 1}), {});
+    const dataImagesCounted = dataImagesArr.reduce((a, b) => Object.assign(a, { [b]: (a[b] || 0) + 1 }), {});
     // an array of all the image values that have duplicates
-    const duplicatedValue = Object.keys(dataImagesCounted).filter((a) => dataImagesCounted[a] > 1)[0];
+    const duplicatedValue = Object.keys(dataImagesCounted).filter(a => dataImagesCounted[a] > 1)[0];
     let duplicateIds = [];
     if (!!duplicatedValue) {
       untypedData.map(row => {
         // send duplicated values to combineRows()
-        if (row.image === duplicatedValue) { duplicateIds.push(row.id); }
+        if (row.image === duplicatedValue) {
+          duplicateIds.push(row.id);
+        }
         if (duplicateIds.length === 2) {
-          const objectKey = `${row.material}-${row.tile}`;
+          const objectKey = `${row.material}-${row.tile.tile}`;
           const row1 = untypedData.filter(obj => obj.id === duplicateIds[0]);
-          const row2 = {[objectKey]: untypedData.filter(obj => obj.id === duplicateIds[1])[0]};
+          const row2 = { [objectKey]: untypedData.filter(obj => obj.id === duplicateIds[1])[0] };
 
           this.combineRows(row1[0], row2);
           const indexToRemove = untypedData.findIndex(x => x.id === row2[objectKey].id);
@@ -82,7 +83,7 @@ export class QuantityService {
           this.updateSummary();
           duplicateIds = [];
         }
-      })
+      });
     }
   }
 
@@ -90,8 +91,9 @@ export class QuantityService {
     this.getRowEstimate(row); // sets feature.estimated_amount
     const editRow = row[Object.keys(row)[0]];
     editRow.total = this.feature.estimated_amount;
-    editRow.tileSqFt = this.getTileSqFt(editRow.tile);
+    editRow.tileSqArea = this.getTileSqArea(editRow.tile);
     editRow.id = this.rowIndexNum++;
+    editRow.material_size = typeof editRow.tile === 'string' ? editRow.tile : editRow.tile.tile;
     this.order.data[index] = editRow;
     this.order.data = this.order.data.slice(); // refreshes the table
     this.updateSummary();
@@ -99,9 +101,15 @@ export class QuantityService {
 
   getRowEstimate(row) {
     switch (this.feature_type) {
-      case 'hush': this.feature.getHushEstimate(row); break;
-      case 'tetria': this.feature.getTetriaEstimate(row); break;
-      case 'clario': this.feature.getClarioEstimate(row); break;
+      case 'hush':
+        this.feature.getHushEstimate(row);
+        break;
+      case 'tetria':
+        this.feature.getTetriaEstimate(row);
+        break;
+      case 'clario':
+        this.feature.getClarioEstimate(row);
+        break;
     }
   }
 
@@ -111,21 +119,26 @@ export class QuantityService {
     let estTotal = 0;
     let tilesUsed = 0;
     let tilesReceiving = 0;
-    let sqFtUsed = 0;
-    let sqFtReceiving = 0;
+    let sqAreaUsed = 0;
+    let sqAreaReceiving = 0;
     summary.map((row: any) => {
       estTotal += row.total;
       tilesUsed += row.used;
       tilesReceiving += row.purchased;
-      sqFtUsed += (row.used * row.tileSqFt);
-      sqFtReceiving  += (row.purchased * row.tileSqFt);
+      sqAreaUsed += row.used * row.tileSqArea;
+      sqAreaReceiving += row.purchased * row.tileSqArea;
     });
     this.estimatedPrice = estTotal;
     this.feature.qtyTilesReceiving = tilesReceiving;
     this.feature.qtyTilesUsed = tilesUsed;
-    this.sqFtUsed = sqFtUsed;
-    this.sqFtReceiving = sqFtReceiving;
-    this.tilesSelected = (sqFtUsed / 4) || null;
+    this.sqAreaUsed = Math.round(sqAreaUsed * 100) / 100;
+    this.sqAreaReceiving = Math.round(sqAreaReceiving * 100) / 100;
+    this.tilesSelected = sqAreaUsed / 4 || null;
+    if (this.feature.feature_type === 'clario' && !!this.clarioGrids.selectedTileSize) {
+      const tileForArea = this.clarioGrids.selectedTileSize.tile_size / 2;
+      const tileArea = this.getTileSqArea(tileForArea.toString());
+      this.tilesSelected = sqAreaUsed / tileArea;
+    }
     this.updateTilesArr();
   }
 
@@ -144,25 +157,41 @@ export class QuantityService {
       const objectKey = `${newObj.material}-${newObj.tile}`;
       if (!tilesArr[objectKey]) {
         tilesArr[objectKey] = newObj;
-      } else { // if tiles are already selected just add to the totals
+      } else {
+        // if tiles are already selected just add to the totals
         tilesArr[objectKey].purchased += newObj.purchased;
         tilesArr[objectKey].used += newObj.used;
       }
-    })
+    });
 
     this.getRowEstimate(tilesArr); // updates feature.ts with the totals
     this.feature.tiles = tilesArr;
   }
 
-  getTileSqFt(tile) {
-    return (tile === '48') ? 8 : 4;
+  getTileSqArea(tile) {
+    switch (tile) {
+      case '24':
+        return 4;
+      case '48':
+        return 8;
+      case '600':
+        return 0.36;
+      case '1200':
+        return 0.36 * 2;
+      case '625':
+        return 0.390625;
+      case '1250':
+        return 0.390625 * 2;
+      default:
+        return 4;
+    }
   }
 }
 
 export interface TileObj {
-  'purchased': number,
-  'image': string,
-  'used': number,
-  'material': string,
-  'tile': string
+  purchased: number;
+  image: string;
+  used: number;
+  material: string;
+  tile: string;
 }
