@@ -7,11 +7,10 @@ import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
   styleUrls: ['./remove-quantity.component.css']
 })
 export class RemoveQuantityComponent implements OnInit {
-
   constructor(
     @Inject(MAT_DIALOG_DATA) public selection: any,
-    private dialogRef: MatDialogRef<RemoveQuantityComponent>,
-  ) { }
+    private dialogRef: MatDialogRef<RemoveQuantityComponent>
+  ) {}
 
   ngOnInit() {}
 
